+++ Editor/Window/ChatElementRunExecutionBlock.cs
@@ -1,0 +1,77 @@
+using Unity.Muse.Agent.Dynamic;
+using Unity.Muse.AppUI.UI;
+using UnityEditor;
+using UnityEngine;
+using UnityEngine.UIElements;
+using UnityEngine.UIElements.Experimental;
+using Button = Unity.Muse.AppUI.UI.Button;
+
+namespace Unity.Muse.Chat
+{
+    class ChatElementRunExecutionBlock : ManagedTemplate
+    {
+        public const string FencedBlockTag = "csx_execute";
+
+        Button m_UndoButton;
+        Text m_Title;
+
+        public ChatElementRunExecutionBlock()
+            : base(MuseChatConstants.UIModulePath)
+        {
+        }
+
+        VisualElement m_ExecutionContainer;
+
+        protected override void InitializeView(TemplateContainer view)
+        {
+            m_Title = view.Q<Text>("actionTitle");
+
+            m_ExecutionContainer = view.Q<VisualElement>("executionContainer");
+
+            m_UndoButton = view.SetupButton("undoButton", _ => UndoHistoryUtils.OpenHistory());
+            m_UndoButton.SetEnabled(false);
+        }
+
+        public void SetData(string data)
+        {
+            if (int.TryParse(data, out var executionId))
+            {
+                var execution = MuseEditorDriver.instance.Agent.RetrieveExecution(executionId);
+
+                m_Title.text = execution.ActionName ?? "Command completed";
+
+                FormatExecutionResult(execution, m_ExecutionContainer);
+
+                m_UndoButton.SetEnabled(execution.SuccessfullyStarted);
+            }
+        }
+
+        public static void FormatExecutionResult(ExecutionResult executionResult, VisualElement container)
+        {
+            if (!executionResult.SuccessfullyStarted)
+            {
+                var executionEntry = new ChatElementRunExecutionEntry(new ExecutionLog($"<color={ExecutionResult.WarningTextColor}>{executionResult.ConsoleLogs}</color>", LogType.Error));
+                executionEntry.Initialize(false);
+                container.Add(executionEntry);
+
+                return;
+            }
+
+            var resultLogs = executionResult.Logs;
+            if (resultLogs.Count == 0)
+            {
+                var executionEntry = new ChatElementRunExecutionEntry(new ExecutionLog("Executed without logs", LogType.Log));
+                executionEntry.Initialize(false);
+                container.Add(executionEntry);
+                return;
+            }
+
+            foreach (var executionLog in resultLogs)
+            {
+                var executionEntry = new ChatElementRunExecutionEntry(executionLog);
+                executionEntry.Initialize(false);
+                container.Add(executionEntry);
+            }
+        }
+    }
+}